Finally fixes that issue where .f would make .faceless not work. or .f fix would make .addclock clock not work.
.f  and .faceless or the newer .remember and .rememberlist previously didn't work due to looping.
check was  - for c in switch; if c in msg - would find all matches where c was in msg (ie for .f: (.f) and (.f)aceless it was contained in,
The msg would get overwritten msg = .faceless, then msg =.f as order of operation.

Now checks book moves (priority) ie .f in .faceless
then check general pbabot moves with a simple if msg in switch (no need for a loop) and would overright the msg from the book (Why else would you put it in the main bot if you didn't want priority can be changed by adding and check ==false

And finally then checks with a for loop (how the clocks are implemented)  so if not msg in switch and msg is empty(from the playbook, though can remove and just if not in pbabot) then for loop and look for clocks or something, last entry in list takes priority for being returned, overwriting other msg.

Also adds .rememberlist and updates .remember stuff.
small tweaks in apoc file.